refactor(server): split task store read paths
