+++ crates/harness-server/src/task_runner/store.rs
@@ -14,6 +14,12 @@ use super::state::{SchedulerAuthorityState, SchedulerOwnerKind, TaskState, TaskS
 use super::types::{TaskId, TaskKind, TaskStatus, TaskTerminalOutcome};
 use super::CompletionCallback;
 
+mod projections;
+mod queries;
+mod runtime_control;
+mod runtime_hosts;
+mod startup;
+
 /// Broadcast channel capacity for per-task stream events.
 /// When the buffer is full, the oldest events are dropped for lagging receivers.
 const TASK_STREAM_CAPACITY: usize = 512;
@@ -73,1094 +79,7 @@ pub struct TaskSummaryPageCursor {
     pub id: String,
 }
 
-impl TaskSummaryFilter {
-    pub fn matches_summary(&self, summary: &TaskSummary) -> bool {
-        if !self.statuses.is_empty() && !self.statuses.contains(&summary.status) {
-            return false;
-        }
-        if self.active && summary.status.is_terminal() {
-            return false;
-        }
-        if !self.scheduler_states.is_empty()
-            && !self
-                .scheduler_states
-                .contains(&summary.scheduler.authority_state)
-        {
-            return false;
-        }
-        if !self.kinds.is_empty() && !self.kinds.contains(&summary.task_kind) {
-            return false;
-        }
-        if self
-            .source
-            .as_deref()
-            .is_some_and(|source| summary.source.as_deref() != Some(source))
-        {
-            return false;
-        }
-        if self
-            .repo
-            .as_deref()
-            .is_some_and(|repo| summary.repo.as_deref() != Some(repo))
-        {
-            return false;
-        }
-        if self
-            .project
-            .as_deref()
-            .is_some_and(|project| summary.project.as_deref() != Some(project))
-        {
-            return false;
-        }
-        true
-    }
-}
-
-fn task_summary_is_after_cursor(summary: &TaskSummary, cursor: &TaskSummaryPageCursor) -> bool {
-    let Some(created_at) = summary
-        .created_at
-        .as_deref()
-        .and_then(|value| DateTime::parse_from_rfc3339(value).ok())
-        .map(|value| value.with_timezone(&Utc))
-    else {
-        return false;
-    };
-    created_at < cursor.created_at
-        || (created_at == cursor.created_at && summary.id.as_str() < cursor.id.as_str())
-}
-
 impl TaskStore {
-    pub async fn open(db_path: &std::path::Path) -> anyhow::Result<Arc<Self>> {
-        Self::open_with_database_url(db_path, None).await
-    }
-
-    pub async fn open_with_database_url(
-        db_path: &std::path::Path,
-        configured_database_url: Option<&str>,
-    ) -> anyhow::Result<Arc<Self>> {
-        let db = TaskDb::open_with_database_url(db_path, configured_database_url).await?;
-        Self::from_task_db(db, db_path).await
-    }
-
-    pub async fn open_with_context(
-        db_path: &std::path::Path,
-        context: &PgStoreContext,
-        setup_pool: &sqlx::postgres::PgPool,
-    ) -> anyhow::Result<Arc<Self>> {
-        let db = TaskDb::open_with_context(context, setup_pool).await?;
-        Self::from_task_db(db, db_path).await
-    }
-
-    pub async fn open_shared_with_data_dir(
-        db_path: &std::path::Path,
-        context: &PgStoreContext,
-        setup_pool: &sqlx::postgres::PgPool,
-        data_dir: &std::path::Path,
-        configured_database_url: Option<&str>,
-    ) -> anyhow::Result<Arc<Self>> {
-        let db = TaskDb::open_shared_with_data_dir(context, setup_pool, data_dir).await?;
-        crate::task_db::migrate_legacy_task_db_if_needed(db_path, configured_database_url, &db)
-            .await?;
-        Self::from_task_db(db, db_path).await
-    }
-
-    pub async fn open_shared_for_reconciliation_with_data_dir(
-        db_path: &std::path::Path,
-        context: &PgStoreContext,
-        setup_pool: &sqlx::postgres::PgPool,
-        data_dir: &std::path::Path,
-        configured_database_url: Option<&str>,
-    ) -> anyhow::Result<Arc<Self>> {
-        let db = TaskDb::open_shared_with_data_dir(context, setup_pool, data_dir).await?;
-        crate::task_db::migrate_legacy_task_db_if_needed(db_path, configured_database_url, &db)
-            .await?;
-        Self::from_task_db_without_startup_recovery(db, db_path).await
-    }
-
-    async fn from_task_db(db: TaskDb, db_path: &std::path::Path) -> anyhow::Result<Arc<Self>> {
-        Self::from_task_db_with_startup_recovery(db, db_path, true).await
-    }
-
-    async fn from_task_db_without_startup_recovery(
-        db: TaskDb,
-        db_path: &std::path::Path,
-    ) -> anyhow::Result<Arc<Self>> {
-        Self::from_task_db_with_startup_recovery(db, db_path, false).await
-    }
-
-    async fn from_task_db_with_startup_recovery(
-        db: TaskDb,
-        db_path: &std::path::Path,
-        run_startup_recovery: bool,
-    ) -> anyhow::Result<Arc<Self>> {
-        // 1. Event replay: runs BEFORE recover_in_progress so event-sourced
-        //    data (pr_url, terminal status) wins over checkpoint data.
-        let event_log_path = db_path
-            .parent()
-            .unwrap_or(std::path::Path::new("."))
-            .join("task-events.jsonl");
-        if run_startup_recovery {
-            if let Err(e) = crate::event_replay::replay_and_recover(&db, &event_log_path).await {
-                tracing::warn!("startup: event replay failed (non-fatal): {e}");
-            }
-
-            // 2. Legacy checkpoint-based recovery as fallback.
-            let recovery = db.recover_in_progress().await?;
-            if recovery.resumed > 0 {
-                tracing::info!(
-                    "startup recovery: resumed {} task(s) from checkpoint",
-                    recovery.resumed
-                );
-            }
-            if recovery.failed > 0 {
-                tracing::warn!(
-                    "startup recovery: marked {} interrupted task(s) as failed (no fresh checkpoint)",
-                    recovery.failed
-                );
-            }
-        }
-
-        // 3. Open the event log for appending during this server session.
-        let event_log = match crate::event_replay::TaskEventLog::open(&event_log_path) {
-            Ok(log) => {
-                tracing::debug!("task event log: {}", event_log_path.display());
-                Some(Arc::new(log))
-            }
-            Err(e) => {
-                tracing::warn!(
-                    "failed to open task event log at {}: {e}",
-                    event_log_path.display()
-                );
-                None
-            }
-        };
-
-        let cache = DashMap::new();
-        let persist_locks = DashMap::new();
-        // Only load active (non-terminal) tasks into the in-memory cache to prevent
-        // unbounded memory growth from historical completed tasks.
-        let active_statuses = &[
-            "pending",
-            "awaiting_deps",
-            "triaging",
-            "planning",
-            "implementing",
-            "review_generating",
-            "review_waiting",
-            "planner_generating",
-            "planner_waiting",
-            "agent_review",
-            "waiting",
-            "reviewing",
-        ];
-        for task in db.list_by_status(active_statuses).await? {
-            persist_locks.insert(task.id.clone(), Arc::new(Mutex::new(())));
-            cache.insert(task.id.clone(), task);
-        }
-        let store = Arc::new(Self {
-            cache,
-            db,
-            persist_locks,
-            stream_txs: DashMap::new(),
-            abort_handles: DashMap::new(),
-            rate_limit_until: RwLock::new(None),
-            event_log,
-        });
-        Ok(store)
-    }
-
-    #[cfg(test)]
-    pub(crate) fn task_db_schema_for_test(&self) -> &str {
-        self.db.schema()
-    }
-
-    #[cfg(test)]
-    pub(crate) fn task_db_store_key_for_test(&self) -> &str {
-        self.db.store_key()
-    }
-
-    pub fn get(&self, id: &TaskId) -> Option<TaskState> {
-        record_task_runner_usage();
-        self.cache.get(id).map(|r| r.value().clone())
-    }
-
-    /// Look up a task by ID, checking the in-memory cache first.
-    /// Falls back to the database for terminal tasks that were evicted from
-    /// the cache at startup (Done, Failed, Cancelled).
-    /// Returns `Ok(None)` only when the ID is unknown in both cache and DB.
-    /// Returns `Err` when the database query itself fails.
-    pub async fn get_with_db_fallback(&self, id: &TaskId) -> anyhow::Result<Option<TaskState>> {
-        record_task_runner_usage();
-        if let Some(task) = self.cache.get(id) {
-            return Ok(Some(task.value().clone()));
-        }
-        self.db.get(id.0.as_str()).await
-    }
-
-    /// Return `true` when the task exists in either the in-memory cache or the
-    /// backing database.
-    ///
-    /// Used by intake reconciliation to detect stale persisted dispatch markers
-    /// after a task-storage backend change (for example SQLite -> Postgres).
-    pub async fn exists_with_db_fallback(&self, id: &TaskId) -> anyhow::Result<bool> {
-        record_task_runner_usage();
-        if self.cache.contains_key(id) {
-            return Ok(true);
-        }
-        self.db.exists_by_id(id.0.as_str()).await
-    }
-
-    /// Return the status of a dependency task with a single DB lookup.
-    ///
-    /// Checks the in-memory cache first; falls back to a lightweight
-    /// `SELECT status` query (no `rounds` JSON decode) for terminal tasks
-    /// evicted from the startup cache.  Returns `None` when the task is
-    /// unknown in both cache and DB, or when the DB call fails.
-    pub(crate) async fn dep_status(&self, id: &TaskId) -> Option<TaskStatus> {
-        if let Some(task) = self.cache.get(id) {
-            return Some(task.status.clone());
-        }
-        match self.db.get_status_only(id.0.as_str()).await {
-            Ok(Some(s)) => s.parse::<TaskStatus>().ok(),
-            Ok(None) => None,
-            Err(e) => {
-                tracing::warn!(task_id = %id.0, "DB error fetching dep status: {e}; treating as absent");
-                None
-            }
-        }
-    }
-
-    /// Return tasks that are in an active status and have not been updated for
-    /// at least `stale_threshold`.  Delegates to [`TaskDb::list_stalled_tasks`].
-    pub async fn list_stalled_tasks(
-        &self,
-        stale_threshold: std::time::Duration,
-        project: Option<&str>,
-    ) -> anyhow::Result<Vec<TaskState>> {
-        record_task_runner_usage();
-        self.db.list_stalled_tasks(stale_threshold, project).await
-    }
-
-    /// Return IDs of terminal tasks (Done, Failed, Cancelled) directly from the database.
-    ///
-    /// Used during startup for worktree cleanup. Only fetches task IDs to avoid
-    /// deserializing the heavy `rounds` column for large historical datasets.
-    pub async fn list_terminal_ids_from_db(&self) -> anyhow::Result<Vec<TaskId>> {
-        record_task_runner_usage();
-        let ids = self
-            .db
-            .list_ids_by_status(&["done", "failed", "cancelled"])
-            .await?;
-        Ok(ids.into_iter().map(harness_core::types::TaskId).collect())
-    }
-
-    pub fn count(&self) -> usize {
-        record_task_runner_usage();
-        self.cache.len()
-    }
-
-    pub(crate) fn postgres_pool(&self) -> sqlx::PgPool {
-        self.db.postgres_pool()
-    }
-
-    #[cfg(test)]
-    pub(crate) fn pool_for_test(&self) -> &sqlx::PgPool {
-        self.db.pool_for_test()
-    }
-
-    /// Check whether an active (non-terminal) task already exists for the same
-    /// project + external_id. Cache-first, DB fallback.
-    pub async fn find_active_duplicate(
-        &self,
-        project_id: &str,
-        external_id: &str,
-    ) -> Option<TaskId> {
-        record_task_runner_usage();
-        let mut found_terminal_in_cache = false;
-        for entry in self.cache.iter() {
-            let task = entry.value();
-            let same_key = task.external_id.as_deref() == Some(external_id)
-                && task
-                    .project_root
-                    .as_ref()
-                    .map(|p| p.to_string_lossy() == project_id)
-                    .unwrap_or(false);
-            if !same_key {
-                continue;
-            }
-            if !matches!(
-                task.status,
-                TaskStatus::Done | TaskStatus::Failed | TaskStatus::Cancelled
-            ) {
-                return Some(task.id.clone());
-            }
-            // Cache has a terminal match — skip DB fallback since cache is
-            // more authoritative and the DB row may be stale.
-            found_terminal_in_cache = true;
-        }
-        if found_terminal_in_cache {
-            return None;
-        }
-        match self.db.find_active_duplicate(project_id, external_id).await {
-            Ok(Some(id)) => Some(harness_core::types::TaskId(id)),
-            Ok(None) => None,
-            Err(e) => {
-                tracing::warn!("dedup: DB lookup failed: {e}");
-                None
-            }
-        }
-    }
-
-    /// Check whether a `done` task with a PR URL already exists for the same
-    /// project + external_id. Cache-first, DB fallback. Fail-open on DB errors.
-    pub async fn find_terminal_pr_duplicate(
-        &self,
-        project_id: &str,
-        external_id: &str,
-    ) -> Option<(TaskId, String)> {
-        record_task_runner_usage();
-        for entry in self.cache.iter() {
-            let task = entry.value();
-            let same_key = task.external_id.as_deref() == Some(external_id)
-                && task
-                    .project_root
-                    .as_ref()
-                    .map(|p| p.to_string_lossy() == project_id)
-                    .unwrap_or(false);
-            if same_key && matches!(task.status, TaskStatus::Done) {
-                if let Some(ref url) = task.pr_url {
-                    return Some((task.id.clone(), url.clone()));
-                }
-            }
-        }
-        match self.db.find_terminal_with_pr(project_id, external_id).await {
-            Ok(Some((id, url))) => Some((harness_core::types::TaskId(id), url)),
-            Ok(None) => None,
-            Err(e) => {
-                tracing::warn!("dedup: terminal PR DB lookup failed: {e}");
-                None
-            }
-        }
-    }
-
-    /// Return all tasks currently in the in-memory cache.
-    ///
-    /// **Semantic note**: since startup only loads active (non-terminal) tasks
-    /// into the cache, this method returns only Pending, AwaitingDeps,
-    /// Implementing, AgentReview, Waiting, and Reviewing tasks.
-    /// To look up a specific completed task use [`get_with_db_fallback`].
-    /// To enumerate all tasks including historical ones use [`list_all_with_terminal`].
-    pub fn list_all(&self) -> Vec<TaskState> {
-        record_task_runner_usage();
-        self.cache.iter().map(|e| e.value().clone()).collect()
-    }
-
-    /// Return all tasks: active ones from cache (most current state) merged
-    /// with terminal tasks from the database (Done, Failed, Cancelled).
-    ///
-    /// Cache entries take precedence over DB rows for the same task ID so that
-    /// in-flight state is always reflected accurately.
-    pub async fn list_all_with_terminal(&self) -> anyhow::Result<Vec<TaskState>> {
-        record_task_runner_usage();
-        // DB is the authoritative record of all tasks ever created.
-        let mut by_id: std::collections::HashMap<TaskId, TaskState> = self
-            .db
-            .list()
-            .await?
-            .into_iter()
-            .map(|t| (t.id.clone(), t))
-            .collect();
-        // Override with cache values — these carry the live in-flight state.
-        for entry in self.cache.iter() {
-            by_id.insert(entry.key().clone(), entry.value().clone());
-        }
-        Ok(by_id.into_values().collect())
-    }
-
-    /// Return all tasks as lightweight [`TaskSummary`] values.
-    ///
-    /// Fetches summary columns from the DB (skipping the heavy `rounds` field),
-    /// then overrides with live cache entries so in-flight state is accurate.
-    /// Use this in the `/tasks` list endpoint instead of `list_all_with_terminal`.
-    pub async fn list_all_summaries_with_terminal(&self) -> anyhow::Result<Vec<TaskSummary>> {
-        record_task_runner_usage();
-        let mut by_id: std::collections::HashMap<TaskId, TaskSummary> = self
-            .db
-            .list_summaries()
-            .await?
-            .into_iter()
-            .map(|s| (s.id.clone(), s))
-            .collect();
-        for entry in self.cache.iter() {
-            by_id.insert(entry.key().clone(), entry.value().summary());
-        }
-        Ok(by_id.into_values().collect())
-    }
-
-    /// Return active tasks as lightweight [`TaskSummary`] values.
-    ///
-    /// Fetches only non-terminal summary rows from the DB, then applies live cache
-    /// overrides so recently terminal tasks are removed and in-flight state wins.
-    pub async fn list_active_summaries(&self) -> anyhow::Result<Vec<TaskSummary>> {
-        record_task_runner_usage();
-        let mut by_id: std::collections::HashMap<TaskId, TaskSummary> = self
-            .db
-            .list_active_summaries()
-            .await?
-            .into_iter()
-            .map(|s| (s.id.clone(), s))
-            .collect();
-        for entry in self.cache.iter() {
-            let summary = entry.value().summary();
-            if summary.status.is_terminal() {
-                by_id.remove(&summary.id);
-            } else {
-                by_id.insert(entry.key().clone(), summary);
-            }
-        }
-        Ok(by_id.into_values().collect())
-    }
-
-    pub async fn list_summaries_filtered_with_terminal(
-        &self,
-        filter: &TaskSummaryFilter,
-    ) -> anyhow::Result<Vec<TaskSummary>> {
-        record_task_runner_usage();
-        let mut by_id: std::collections::HashMap<TaskId, TaskSummary> = self
-            .db
-            .list_summaries_filtered(filter)
-            .await?
-            .into_iter()
-            .map(|summary| (summary.id.clone(), summary))
-            .collect();
-        for entry in self.cache.iter() {
-            let summary = entry.value().summary();
-            if filter.matches_summary(&summary) {
-                by_id.insert(entry.key().clone(), summary);
-            } else {
-                by_id.remove(entry.key());
-            }
-        }
-        Ok(by_id.into_values().collect())
-    }
-
-    pub async fn list_summaries_filtered_page_with_terminal(
-        &self,
-        filter: &TaskSummaryFilter,
-        cursor: Option<&TaskSummaryPageCursor>,
-        limit: usize,
-    ) -> anyhow::Result<Vec<TaskSummary>> {
-        record_task_runner_usage();
-        let db_limit = limit.saturating_add(self.cache.len()).saturating_add(1);
-        let mut by_id: std::collections::HashMap<TaskId, TaskSummary> = self
-            .db
-            .list_summaries_filtered_page(filter, cursor, db_limit)
-            .await?
-            .into_iter()
-            .map(|summary| (summary.id.clone(), summary))
-            .collect();
-        for entry in self.cache.iter() {
-            let summary = entry.value().summary();
-            if filter.matches_summary(&summary)
-                && cursor.is_none_or(|cursor| task_summary_is_after_cursor(&summary, cursor))
-            {
-                by_id.insert(entry.key().clone(), summary);
-            } else {
-                by_id.remove(entry.key());
-            }
-        }
-        Ok(by_id.into_values().collect())
-    }
-
-    /// Return `(TaskId, TaskStatus)` pairs for all tasks without deserializing `rounds`.
-    ///
-    /// Hot-path callers (skill governance, token usage attribution) that only need
-    /// task status should use this instead of `list_all_with_terminal`.
-    pub async fn list_all_statuses_with_terminal(
-        &self,
-    ) -> anyhow::Result<HashMap<TaskId, TaskStatus>> {
-        record_task_runner_usage();
-        let mut by_id: HashMap<TaskId, TaskStatus> = self
-            .db
-            .list_id_status()
-            .await?
-            .into_iter()
-            .filter_map(|(id, status)| {
-                status
-                    .parse::<TaskStatus>()
-                    .ok()
-                    .map(|s| (harness_core::types::TaskId(id), s))
-            })
-            .collect();
-        for entry in self.cache.iter() {
-            by_id.insert(entry.key().clone(), entry.value().status.clone());
-        }
-        Ok(by_id)
-    }
-
-    /// Return the latest status for each `external_id` in a single project/repo namespace.
-    ///
-    /// The database query collapses historical attempts down to the newest row per
-    /// issue. In-memory cache entries then override the DB result when the server
-    /// is holding a newer status transition that has not been re-read from storage.
-    pub async fn list_latest_external_statuses_by_project_and_repo(
-        &self,
-        project_id: &str,
-        repo: Option<&str>,
-    ) -> anyhow::Result<HashMap<String, (Option<String>, TaskStatus)>> {
-        record_task_runner_usage();
-        let mut by_external_id: HashMap<String, (Option<String>, TaskStatus)> = self
-            .db
-            .list_latest_external_statuses_by_project_and_repo(project_id, repo)
-            .await?
-            .into_iter()
-            .filter_map(|(external_id, status, created_at)| {
-                status
-                    .parse::<TaskStatus>()
-                    .ok()
-                    .map(|parsed| (external_id, (created_at, parsed)))
-            })
-            .collect();
-        for entry in self.cache.iter() {
-            let task = entry.value();
-            let same_project = task
-                .project_root
-                .as_ref()
-                .map(|path| path.to_string_lossy() == project_id)
-                .unwrap_or(false);
-            let same_repo = match (repo, task.repo.as_deref()) {
-                (Some(expected), Some(actual)) => expected == actual,
-                (None, None) => true,
-                _ => false,
-            };
-            if !same_project || !same_repo {
-                continue;
-            }
-            let Some(external_id) = task.external_id.clone() else {
-                continue;
-            };
-            let should_replace = by_external_id
-                .get(&external_id)
-                .map(|(existing_created_at, _)| {
-                    latest_timestamp_at_least(
-                        task.created_at.as_deref(),
-                        existing_created_at.as_deref(),
-                    )
-                })
-                .unwrap_or(true);
-            if should_replace {
-                by_external_id.insert(external_id, (task.created_at.clone(), task.status.clone()));
-            }
-        }
-
-        Ok(by_external_id)
-    }
-
-    /// Count live runtime-host leases from the authoritative scheduler state
-    /// carried on active task rows.
-    pub(crate) fn active_runtime_host_lease_count(&self, host_id: &str) -> usize {
-        let now = Utc::now();
-        self.cache
-            .iter()
-            .filter(|entry| {
-                let scheduler = &entry.value().scheduler;
-                scheduler.runtime_host_id() == Some(host_id)
-                    && scheduler.has_live_runtime_host_lease(now)
-            })
-            .count()
-    }
-
-    /// Attempt to claim a pending task for a runtime host by updating the
-    /// authoritative scheduler state on the task row itself.
-    pub(crate) async fn claim_for_runtime_host(
-        &self,
-        id: &TaskId,
-        host_id: &str,
-        lease_secs: Option<i64>,
-    ) -> anyhow::Result<Option<crate::runtime_hosts::TaskClaimResult>> {
-        let ttl = lease_secs
-            .unwrap_or(crate::runtime_hosts::DEFAULT_LEASE_SECS)
-            .max(0);
-        let lease_duration = chrono::TimeDelta::try_seconds(ttl).ok_or_else(|| {
-            anyhow::anyhow!(
-                "lease_secs value {ttl} is too large to compute a valid expiration timestamp"
-            )
-        })?;
-        let now = Utc::now();
-        let expires_at = now.checked_add_signed(lease_duration).ok_or_else(|| {
-            anyhow::anyhow!(
-                "lease_secs value {ttl} is too large to compute a valid expiration timestamp"
-            )
-        })?;
-        let lock = self
-            .persist_locks
-            .entry(id.clone())
-            .or_insert_with(|| Arc::new(Mutex::new(())))
-            .clone();
-        let _guard = lock.lock().await;
-
-        let Some(mut entry) = self.cache.get_mut(id) else {
-            return Ok(None);
-        };
-        if !matches!(entry.status, TaskStatus::Pending) {
-            return Ok(None);
-        }
-        let original_scheduler = entry.scheduler.clone();
-        if let Some(owner) = entry.scheduler.owner.as_ref() {
-            match owner.kind {
-                SchedulerOwnerKind::Scheduler => {
-                    // Pending rows owned by the local scheduler are unreachable today,
-                    // but if one ever appears we must not let a runtime host steal it.
-                    return Ok(None);
-                }
-                SchedulerOwnerKind::RuntimeHost => {
-                    // RuntimeHost with live lease: the claim is still active.
-                    if entry.scheduler.has_live_runtime_host_lease(now) {
-                        return Ok(None);
-                    }
-                    // RuntimeHost with a stale lease: safe to clear and re-claim.
-                    entry.scheduler.clear_to_queued();
-                }
-            }
-        }
-
-        entry.scheduler.claim_runtime_host(host_id, expires_at);
-        let snapshot = entry.value().clone();
-        drop(entry);
-        if let Err(e) = self.db.update(&snapshot).await {
-            if let Some(mut rollback) = self.cache.get_mut(id) {
-                rollback.scheduler = original_scheduler;
-            }
-            return Err(e);
-        }
-        if let Some(mut entry) = self.cache.get_mut(id) {
-            entry.value_mut().version = entry.value().version.saturating_add(1);
-        }
-        Ok(Some(crate::runtime_hosts::TaskClaimResult {
-            task_id: id.clone(),
-            lease_expires_at: expires_at.to_rfc3339(),
-        }))
-    }
-
-    /// Clear a pending runtime-host lease from the authoritative scheduler state.
-    ///
-    /// Returns `true` only when the task still belonged to `host_id` and was
-    /// rewritten back to the queued state.
-    pub(crate) async fn release_runtime_host_claim(
-        &self,
-        id: &TaskId,
-        host_id: &str,
-    ) -> anyhow::Result<bool> {
-        let lock = self
-            .persist_locks
-            .entry(id.clone())
-            .or_insert_with(|| Arc::new(Mutex::new(())))
-            .clone();
-        let _guard = lock.lock().await;
-
-        let Some(mut entry) = self.cache.get_mut(id) else {
-            return Ok(false);
-        };
-        if !matches!(entry.status, TaskStatus::Pending) {
-            return Ok(false);
-        }
-        if entry.scheduler.runtime_host_id() != Some(host_id) {
-            return Ok(false);
-        }
-
-        let original_scheduler = entry.scheduler.clone();
-        entry.scheduler.clear_to_queued();
-        let snapshot = entry.value().clone();
-        drop(entry);
-        if let Err(e) = self.db.update(&snapshot).await {
-            if let Some(mut rollback) = self.cache.get_mut(id) {
-                rollback.scheduler = original_scheduler;
-            }
-            return Err(e);
-        }
-        if let Some(mut entry) = self.cache.get_mut(id) {
-            entry.value_mut().version = entry.value().version.saturating_add(1);
-        }
-        Ok(true)
-    }
-
-    /// Release every pending task lease owned by `host_id`.
-    pub(crate) async fn release_runtime_host_claims(
-        &self,
-        host_id: &str,
-    ) -> anyhow::Result<Vec<TaskId>> {
-        let candidate_ids: Vec<TaskId> = self
-            .cache
-            .iter()
-            .filter_map(|entry| {
-                let task = entry.value();
-                (matches!(task.status, TaskStatus::Pending)
-                    && task.scheduler.runtime_host_id() == Some(host_id))
-                .then(|| task.id.clone())
-            })
-            .collect();
-
-        let mut released = Vec::new();
-        for task_id in candidate_ids {
-            if self.release_runtime_host_claim(&task_id, host_id).await? {
-                released.push(task_id);
-            }
-        }
-        Ok(released)
-    }
-
-    /// Return the `pr_url` of the most recently created Done task, ordered by `created_at DESC`
-    /// from the database (stable ordering, unlike the in-memory DashMap cache).
-    pub async fn latest_done_pr_url(&self) -> Option<String> {
-        record_task_runner_usage();
-        match self.db.latest_done_pr_url().await {
-            Ok(url) => url,
-            Err(e) => {
-                tracing::warn!("failed to query latest done PR URL: {e}");
-                None
-            }
-        }
-    }
-
-    /// Return the `pr_url` of the most recent Done task for a specific project root path.
-    pub async fn latest_done_pr_url_by_project(&self, project: &str) -> Option<String> {
-        record_task_runner_usage();
-        match self.db.latest_done_pr_url_by_project(project).await {
-            Ok(url) => url,
-            Err(e) => {
-                tracing::warn!("failed to query latest done PR URL for project {project}: {e}");
-                None
-            }
-        }
-    }
-
-    /// Fetch the latest done PR URL for every project in a single DB query.
-    /// Returns a map from project root path string to PR URL.
-    pub async fn latest_done_pr_urls_all_projects(&self) -> HashMap<String, String> {
-        record_task_runner_usage();
-        match self.db.latest_done_pr_urls_all_projects().await {
-            Ok(map) => map,
-            Err(e) => {
-                tracing::warn!("failed to bulk-query latest done PR URLs: {e}");
-                HashMap::new()
-            }
-        }
-    }
-
-    /// Return all inflight sibling tasks on the same project, excluding `exclude_id`.
-    ///
-    /// Used by `run_task` to build sibling-awareness context before starting implementation,
-    /// so each agent knows what other agents are working on and avoids touching their files.
-    /// Only tasks that have `project_root` set (populated at spawn time) are considered.
-    pub fn list_siblings(&self, project: &std::path::Path, exclude_id: &TaskId) -> Vec<TaskState> {
-        record_task_runner_usage();
-        self.cache
-            .iter()
-            .filter(|e| {
-                let task = e.value();
-                &task.id != exclude_id
-                    && (matches!(task.status, TaskStatus::Pending) || task.status.is_inflight())
-                    && task.project_root.as_deref() == Some(project)
-            })
-            .map(|e| e.value().clone())
-            .collect()
-    }
-
-    /// Compute global and per-project done/failed counts via SQL aggregation.
-    ///
-    /// Delegates to `TaskDb` so the count scales with the database engine rather
-    /// than requiring an O(N) scan of the in-memory cache, which grows unboundedly
-    /// as tasks accumulate. Uses the `idx_tasks_project_status_updated` index.
-    pub async fn count_for_dashboard(&self) -> DashboardCounts {
-        record_task_runner_usage();
-        match self.db.count_done_failed_by_project().await {
-            Ok((global_done, global_failed, global_stalled, rows)) => {
-                let by_project = rows
-                    .into_iter()
-                    .map(|(project, done, failed, stalled)| {
-                        (
-                            project,
-                            ProjectCounts {
-                                done,
-                                failed,
-                                stalled,
-                            },
-                        )
-                    })
-                    .collect();
-                DashboardCounts {
-                    global_done,
-                    global_failed,
-                    global_stalled,
-                    by_project,
-                }
-            }
-            Err(e) => {
-                tracing::warn!("count_for_dashboard: SQL aggregation failed: {e}");
-                DashboardCounts {
-                    global_done: 0,
-                    global_failed: 0,
-                    global_stalled: 0,
-                    by_project: HashMap::new(),
-                }
-            }
-        }
-    }
-
-    /// Count tasks that completed (`done`) with `updated_at >= since`.
-    ///
-    /// Forwards to [`TaskDb::count_done_since`]; used by the system overview
-    /// to compute merged-in-last-24h without materialising full task rows.
-    pub async fn count_done_since(&self, since: chrono::DateTime<chrono::Utc>) -> u64 {
-        record_task_runner_usage();
-        match self.db.count_done_since(since).await {
-            Ok(v) => v,
-            Err(e) => {
-                tracing::warn!("count_done_since: SQL aggregation failed: {e}");
-                0
-            }
-        }
-    }
-
-    /// Per-(project, hour) done counts since `since`. Forwards to
-    /// [`TaskDb::done_per_project_hour_since`] and returns an empty vec on
-    /// error so callers can degrade gracefully.
-    pub async fn done_per_project_hour_since(
-        &self,
-        since: chrono::DateTime<chrono::Utc>,
-    ) -> Vec<(String, String, u64)> {
-        record_task_runner_usage();
-        match self.db.done_per_project_hour_since(since).await {
-            Ok(rows) => rows,
-            Err(e) => {
-                tracing::warn!("done_per_project_hour_since: SQL aggregation failed: {e}");
-                Vec::new()
-            }
-        }
-    }
-
-    /// Collect lightweight inputs for LLM metrics computation.
-    ///
-    /// Both **turn counts** and **first-token latencies** are collected in two
-    /// bounded phases so the dashboard poll remains O(1) regardless of total
-    /// task history size:
-    ///
-    /// 1. **Cache phase** — iterate `DashMap` refs in-place (no full `TaskState`
-    ///    clone).  Records the IDs seen so phase 2 can deduplicate.
-    /// 2. **DB phase** — bounded SQL queries (`LIMIT 500`, `ORDER BY updated_at DESC`)
-    ///    return only scalar columns (`turn`, `first_token_latency_ms`) for the
-    ///    500 most-recently-completed terminal tasks not already counted from cache.
-    ///    This covers the idle-queue restart case without an unbounded table scan.
-    ///
-    /// Rounds whose `result` is `"resumed_checkpoint"` are skipped in both
-    /// sources: they are synthetic markers with no real latency data.
-    pub async fn collect_llm_metrics_inputs(&self) -> LlmMetricsInputs {
-        record_task_runner_usage();
-        // Phase 1: iterate cache refs in-place; never clone full TaskState.
-        // Collect both turn counts and latencies in a single pass, and track
-        // IDs seen so we can deduplicate against the DB queries in phase 2.
-        //
-        // Only terminal (done/failed) tasks are counted so the cache phase matches
-        // DB phase semantics: both sources use the same status set.  Non-terminal
-        // task IDs are still inserted into `cache_ids` so the DB deduplication
-        // logic remains correct (we never double-count a task that transitions to
-        // terminal while the DB query is in-flight).
-        let mut cache_ids: std::collections::HashSet<String> = std::collections::HashSet::new();
-        let mut turn_counts: Vec<u32> = Vec::new();
-        let mut first_token_latencies: Vec<u64> = Vec::new();
-        for entry in self.cache.iter() {
-            let task = entry.value();
-            cache_ids.insert(entry.key().0.clone());
-
-            // Skip non-terminal tasks to match the DB phase (done/failed only).
-            if !matches!(task.status, TaskStatus::Done | TaskStatus::Failed) {
-                continue;
-            }
-
-            if task.turn > 0 {
-                turn_counts.push(task.turn);
-            }
-
-            // Skip the synthetic resumed_checkpoint round injected at recovery time,
-            // then find the first round that actually received a response token.
-            // Using find_map (instead of find + and_then) means transient_retry rounds
-            // whose first_token_latency_ms is None are also skipped, so a task with
-            // one or more transient failures before a successful round is still included
-            // in p50 using the latency of that successful round.
-            if let Some(latency) = task
-                .rounds
-                .iter()
-                .filter(|r| r.result != "resumed_checkpoint")
-                .find_map(|r| r.first_token_latency_ms)
-            {
-                first_token_latencies.push(latency);
-            }
-        }
-
-        // Phase 2a: bounded DB query for terminal turn counts not in cache.
-        // After restart with idle queue the cache is empty; without this, turn
-        // counts would be zero even though history is persisted in the DB.
-        // `list_terminal_turn_counts` fetches only `id` and `turn` (no rounds
-        // blob) and is bounded to 500 rows ordered by updated_at DESC.
-        match self.db.list_terminal_turn_counts().await {
-            Ok(rows) => {
-                for (id, turn) in rows {
-                    if !cache_ids.contains(&id) && turn > 0 {
-                        turn_counts.push(turn as u32);
-                    }
-                }
-            }
-            Err(e) => {
-                tracing::warn!("collect_llm_metrics_inputs: DB turn counts query failed: {e}");
-            }
-        }
-
-        // Phase 2b: bounded DB query for terminal latencies not in cache.
-        // `list_terminal_first_token_latencies_ms` extracts the scalar in SQL via
-        // json_extract (no full rounds blob in Rust memory) and is bounded to the
-        // 500 most-recently-completed terminal tasks so each dashboard poll stays O(1).
-        match self.db.list_terminal_first_token_latencies_ms().await {
-            Ok(rows) => {
-                for (id, latency_opt) in rows {
-                    if cache_ids.contains(&id) {
-                        // Already counted via cache iteration above.
-                        continue;
-                    }
-                    if let Some(ms) = latency_opt {
-                        first_token_latencies.push(ms as u64);
-                    }
-                }
-            }
-            Err(e) => {
-                tracing::warn!("collect_llm_metrics_inputs: DB terminal latency query failed: {e}");
-            }
-        }
-
-        LlmMetricsInputs {
-            turn_counts,
-            first_token_latencies,
-        }
-    }
-
-    /// Return all cached tasks whose `parent_id` matches the given ID.
-    /// Reconstructs the child list from in-memory state; does not require
-    /// `subtask_ids` to be persisted on the parent.
-    pub fn list_children(&self, parent_id: &TaskId) -> Vec<TaskState> {
-        record_task_runner_usage();
-        self.cache
-            .iter()
-            .filter(|e| e.value().parent_id.as_ref() == Some(parent_id))
-            .map(|e| e.value().clone())
-            .collect()
-    }
-
-    /// Register a broadcast channel for a task's stream. Call once before task execution starts.
-    pub(crate) fn register_task_stream(&self, id: &TaskId) {
-        let (tx, _rx) = broadcast::channel(TASK_STREAM_CAPACITY);
-        self.stream_txs.insert(id.clone(), tx);
-    }
-
-    /// Subscribe to a task's active stream. Returns `None` if no stream is registered.
-    pub fn subscribe_task_stream(&self, id: &TaskId) -> Option<broadcast::Receiver<StreamItem>> {
-        record_task_runner_usage();
-        self.stream_txs.get(id).map(|tx| tx.subscribe())
-    }
-
-    /// Publish a [`StreamItem`] to all current subscribers of a task stream.
-    /// No-op when no stream is registered. Dropping the send result is intentional:
-    /// `SendError` only occurs when there are no active receivers, which is normal
-    /// when no SSE client is connected (backpressure: oldest events dropped on lag).
-    pub(crate) fn publish_stream_item(&self, id: &TaskId, item: StreamItem) {
-        if let Some(tx) = self.stream_txs.get(id) {
-            if let Err(e) = tx.send(item) {
-                tracing::trace!(task_id = %id.0, "stream publish dropped (no receivers): {e}");
-            }
-        }
-    }
-
-    /// Remove the task's stream channel after execution completes.
-    pub(crate) fn close_task_stream(&self, id: &TaskId) {
-        self.stream_txs.remove(id);
-    }
-
-    /// Store the abort handle for a running task so it can be cancelled later.
-    pub(crate) fn store_abort_handle(&self, id: &TaskId, handle: tokio::task::AbortHandle) {
-        self.abort_handles.insert(id.clone(), handle);
-    }
-
-    /// Remove and discard the abort handle when the task finishes normally.
-    pub(crate) fn remove_abort_handle(&self, id: &TaskId) {
-        self.abort_handles.remove(id);
-    }
-
-    /// Abort the running task's Tokio future, if one is registered.
-    /// The `kill_on_drop(true)` flag on child processes ensures the CLI subprocess
-    /// is also killed when the future is dropped after abort.
-    /// Returns `true` if an abort handle was found and triggered.
-    pub fn abort_task(&self, id: &TaskId) -> bool {
-        record_task_runner_usage();
-        if let Some(handle) = self.abort_handles.get(id) {
-            handle.abort();
-            true
-        } else {
-            false
-        }
-    }
-
-    /// Activate the global rate-limit circuit breaker. All tasks will pause
-    /// before their next agent call until `duration` elapses.
-    pub async fn set_rate_limit(&self, duration: std::time::Duration) {
-        record_task_runner_usage();
-        let until = tokio::time::Instant::now() + duration;
-        *self.rate_limit_until.write().await = Some(until);
-        tracing::warn!(
-            pause_secs = duration.as_secs(),
-            "rate-limit circuit breaker activated — all tasks paused"
-        );
-    }
-
-    /// Wait if the global rate-limit circuit breaker is active.
-    /// Returns only after the deadline is None or already past. Loops to handle
-    /// the case where a concurrent `set_rate_limit()` extends the deadline while
-    /// this task is sleeping — without looping, the task would proceed after the
-    /// original deadline and burn turns / trigger extra 429s.
-    pub async fn wait_for_rate_limit(&self) {
-        record_task_runner_usage();
-        loop {
-            let deadline = { *self.rate_limit_until.read().await };
-            let Some(until) = deadline else { return };
-            let now = tokio::time::Instant::now();
-            if now >= until {
-                // Deadline already passed. Clear it if unchanged; if a concurrent
-                // setter extended the deadline, re-loop to wait for the new value.
-                let mut wl = self.rate_limit_until.write().await;
-                if *wl == Some(until) {
-                    *wl = None;
-                    return;
-                }
-                // Deadline was extended concurrently — fall through to re-loop.
-                continue;
-            }
-            let remaining = until - now;
-            tracing::info!(
-                remaining_secs = remaining.as_secs(),
-                "task waiting for rate-limit circuit breaker to clear"
-            );
-            tokio::time::sleep_until(until).await;
-            // Only clear the breaker if it still holds the same deadline we
-            // snapshotted. A concurrent set_rate_limit() call may have extended
-            // the deadline during our sleep; preserving the new deadline lets
-            // the next loop iteration wait for the extension.
-            {
-                let mut wl = self.rate_limit_until.write().await;
-                if *wl == Some(until) {
-                    *wl = None;
-                }
-            }
-            // Loop again: re-read the deadline in case it was extended.
-        }
-    }
-
     /// Persist an artifact captured from agent output during task execution.
     pub(crate) async fn insert_artifact(
         &self,
